move extract kinds to a specific directory
