+++ src/classifier/extractKindsFromTs2345ErrorMessage.ts
